requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
dtaidistance>=2.3.10
tensorboard>=2.14.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

